requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
Brotli==1.1.0
browser-cookie3==0.19.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
curl-cffi==0.5.9
dnspython==2.4.2
duckduckgo-search==3.9.4
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
g4f==0.1.7.9
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
Js2Py==0.74
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.3
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.1
orjson==3.9.10
Pillow==10.1.0
platformdirs==3.11.0
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyExecJS==1.5.1
pyjsparser==2.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
starlette==0.27.0
tiktoken==0.5.1
typing_extensions==4.8.0
tzlocal==5.2
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.19.0
waitress==2.1.2
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
